+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.3.2
+Changes:
+* Added contribution guideline, Makefile and other developer experience improvements
+* Started raising `ConfigurationError` if a `Model` attribute is an explicit `Field` but lacks a type annotation
+
 ## v1.3.1
 Fixes:
 * Fixed `Missing` behavior with `deepcopy`